Move UIKitInstrumentedTests.sh script to the repo from CI (#2115)
Experiment to fix
https://youtrack.jetbrains.com/issue/CMP-8136/PR-checks-runs-always-by-main-teamcity-config

The code is moved from
https://jetbrains.team/p/ui/repositories/compose-teamcity-config/files/781b140e33a142c7a3428a9cfebfcdb37bf96723/.teamcity/compose/pullRequests/UIKitInstrumentedTests.kt?tab=source&line=36&lines-count=1

If it is successful, later other PR checks will be written this way.

CI will call it this way:
```
[ -f .jb-ciscripts/PullRequests/UIKitInstrumentedTests.sh ] && .jb-ciscripts/PullRequests/UIKitInstrumentedTests.sh
```

In the release branch it doesn't exist, so it will be green. In the main
branch it will run all the tests

## Release Notes
N/A