Add linuxMain source set (#2007)
This PR creates a new `linuxMain` source set which depends on
`nativeMain`. It can be created with `linux()`, this is similar to how
it is done for other targets.

Replaced individual `linuxX64()`/`linuxArm64()` targets with this new
`linux()` call. I did not add linux support to any new modules, this PR
is so the actual PR supporting Linux for Compose UI will be a bit
smaller. The additional linuxMain source set is required for that.

## Testing
N/A

## Release Notes
N/A